python/requirements.txt
uiautomator2
appium-python-client
selenium